requirements.txt
# AI Signal — Python dependencies

# Google Gemini SDK (free tier — no credit card required)
# Get your key at: https://aistudio.google.com/apikey
google-generativeai>=0.8.0

# RSS/Atom feed parsing (Python CLI)
feedparser>=6.0.11

# Environment variable loading
python-dotenv>=1.0.0

# Production server (optional — for server/proxy.py)
flask>=3.0.0
gunicorn>=21.0.0
